+++ src/main/java/org/ftc9974/thorcore/samples/SampleEncoders.java
@@ -18,7 +18,6 @@ import java.util.Locale;
  * sample may make the robot turn in place instead of driving forward.
  */
 @TeleOp(name = "Sample - Encoders", group = "ThorCore Samples")
-@Disabled
 public class SampleEncoders extends LinearOpMode {
 
     private TankDrive4Wheel rb;