+++ src/main/java/org/ftc9974/thorcore/samples/SampleNEONStackVision.java
@@ -1,5 +1,6 @@
 package org.ftc9974.thorcore.samples;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.RobotLog;