+++ src/main/java/org/ftc9974/thorcore/samples/SamplePixy2USB.java
@@ -26,7 +26,6 @@ import java.util.List;
  * @see Pixy2USB
  */
 @TeleOp(name = "Sample - Pixy2USB", group = "ThorCore Samples")
-//@Disabled
 public class SamplePixy2USB extends OpMode {
 
     private Pixy2USB pixy;